+++ php/user.php
@@ -76,7 +76,7 @@
                 <?php if(isset($_SESSION['username'])){
                     echo '
                     <div>
-                        <a href="notification.php" id="notification">
+                        <a href="notificationlist.php" id="notification">
                             <span class="glyphicon">&#x2709;</span>
                         </a>
                         <a class="navbar-brand" style="transition: 0.5s;" id="user-page-active" href="user.php?username=' . $_SESSION['username'] . '">' . $_SESSION['username'] . '